requirement.txt
Flask==1.1.2
requests==2.23.0
slackclient==2.5.0
slackeventsapi==2.1.0
urllib3==1.25.8

